+++ packages/preview-middleware-client/src/adp/controllers/AddTableColumnFragments.controller.ts
@@ -283,8 +283,8 @@ export default class AddTableColumnFragments extends BaseDialog<AddTableColumnsF
                 targetObject,
                 'addXML',
                 modifiedValue,
-                designMetadata,
-                flexSettings
+                flexSettings,
+                designMetadata
             );
 
             const templateName =