+++ packages/preview-middleware-client/test/unit/adp/init.test.ts
@@ -11,6 +11,16 @@ import type RuntimeAuthoring from 'sap/ui/rta/RuntimeAuthoring';
 import ElementRegistry from 'mock/sap/ui/core/ElementRegistry';
 import Element from 'mock/sap/ui/core/Element';
 
+const addFragmentServiceMock = jest.fn();
+jest.mock('open/ux/preview/client/adp/add-fragment', () => ({
+    initAddXMLPlugin: addFragmentServiceMock
+}));
+
+const extendControllerServiceMock = jest.fn();
+jest.mock('open/ux/preview/client/adp/extend-controller', () => ({
+    initExtendControllerPlugin: extendControllerServiceMock
+}));
+
 describe('adp', () => {
     const addMenuItemSpy = jest.fn();
     let initOutlineSpy: jest.SpyInstance;
@@ -164,5 +174,24 @@ describe('adp', () => {
                 shouldHideIframe: false
             }
         });
+    });
+
+    test('init - use AddXMLPlugin and ExtendControllerPlugin for UI5 version higher than 1.136', async () => {
+        const mockUI5Element = {
+            getMetadata: jest.fn().mockReturnValue({
+                getName: jest.fn().mockReturnValue('XMLView')
+            }),
+            oAsyncState: undefined,
+            getId: jest.fn().mockReturnValue('application-app-preview-component---fin.ar.lineitems.display.appView')
+        };
+
+        Element.registry.filter.mockReturnValue([mockUI5Element]);
+
+        VersionInfo.load.mockResolvedValue({ name: 'sap.ui.core', version: '1.137.0' });
+
+        await init(rtaMock as unknown as RuntimeAuthoring);
+
+        expect(addFragmentServiceMock).toHaveBeenCalledWith(rtaMock)
+        expect(extendControllerServiceMock).toHaveBeenCalledWith(rtaMock);
     });
 });