+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/project-access
 
+## 1.30.1
+
+### Patch Changes
+
+-   Updated dependencies [5e0020b]
+    -   @sap-ux/ui5-config@0.28.1
+
+## 1.30.0
+
+### Minor Changes
+
+-   a28357d: chore - drop node18 support as it is out of maintenance
+
+### Patch Changes
+
+-   Updated dependencies [a28357d]
+    -   @sap-ux/ui5-config@0.28.0
+    -   @sap-ux/i18n@0.3.0
+
 ## 1.29.22
 
 ### Patch Changes